[FIXED] Project Scripts and structure
