+++ package.json
@@ -11,7 +11,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-loading": "^2.0.3",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^2.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
