+++ src/components/Loading.jsx
@@ -1,9 +1,10 @@
+import React from 'react';
 import ReactLoading from 'react-loading';
-const Carregando = ({ type, color, height, width }) => (
+const Loading = ({ type, color, height, width }) => (
     <div>
         <ReactLoading type={type} color={color} height={height} width={width} />
 
     </div>
 );
  
-export default Carregando;
+export default Loading;
